Modified the API to use Object instead of BigInteger, allowing someone to write a partition without having to use BigInteger everywhere. This is a first step, we should also get rid of BigInteger completly, to use long instead.
git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@517022 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/directory/server/core/partition/impl/btree/BTreeSearchResultEnumeration.java
@@ -20,6 +20,7 @@
 package org.apache.directory.server.core.partition.impl.btree;
 
 
+import java.math.BigInteger;
 import java.util.Iterator;
 
 import javax.naming.NamingEnumeration;
@@ -104,11 +105,11 @@ public class BTreeSearchResultEnumeration implements SearchResultEnumeration
     {
         IndexRecord rec = ( IndexRecord ) underlying.next();
         Attributes entry;
-        String name = partition.getEntryUpdn( rec.getEntryId() );
+        String name = partition.getEntryUpdn( (BigInteger)rec.getEntryId() );
 
         if ( null == rec.getAttributes() )
         {
-            rec.setAttributes( partition.lookup( rec.getEntryId() ) );
+            rec.setAttributes( partition.lookup( (BigInteger)rec.getEntryId() ) );
         }
 
         if ( attrIds == null )
@@ -234,7 +235,7 @@ public class BTreeSearchResultEnumeration implements SearchResultEnumeration
             }
         }
 
-        BTreeSearchResult result = new BTreeSearchResult( rec.getEntryId(), name, null, entry );
+        BTreeSearchResult result = new BTreeSearchResult( (BigInteger)rec.getEntryId(), name, null, entry );
         result.setRelative( false );
         return result;
     }
